--- ControlBarra.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package lluvia_de_estrellas;
-
-/**
- *
- * @author guille
- */
-public class ControlBarra{
-    public ControlBarra(){
-        System.out.println("Prueba");
-    }
-    
-    
-}
