--- Game.java
@@ -1,26 +1,0 @@
-/*
- * 
- * By  Adrian Garcia San Jose.
- * 
- */
-package lluvia_de_estrellas;
-
-/**
- *
- * @author adri
- */
-public class Game {
-    
-    /*
-      Jorge prueva pull
-    */
-    
-   
-    Vista vista=new Vista();
-    
-    ControlMenu controlMenu=new ControlMenu(vista);
- 
-    
-    
-    
-}
