--- Lluvia_de_estrellas.java
@@ -1,21 +1,0 @@
-/*
- * 
- * By  Adrian Garcia San Jose.
- * 
- */
-package lluvia_de_estrellas;
-
-/**
- *
- * @author adri
- */
-public class Lluvia_de_estrellas {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        Game juego=new Game();
-    }
-    
-}
